+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeteredData/V1/Model/ForwardMeteredDataCommandV1.cs
@@ -19,7 +19,7 @@ using Energinet.DataHub.ProcessManager.Abstractions.Client;
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
 
 /// <summary>
-/// Command for starting a BRS-021-ForwardMeteredData.
+/// Command for starting a BRS-021 forward metered data.
 /// Must be JSON serializable.
 /// </summary>
 public record ForwardMeteredDataCommandV1